refactor(example): murder - background
