+++ example/vite-demo/murder/engine/background.ts
